Make years_control an integer and adjust UI
Use an integer years_control throughout the stack and update related UI/loader code. trade_value.py now writes the integer `n` for years_control (instead of a rounded fractional), the Player model column was changed to Integer, and PLAYER_INT_COLS in data_loader includes years_control so it's parsed as an int. Frontend changes: TradeValueChart now rounds hovered yearsControl and PlayerDetails uses yrsCtrl directly (removed floor) when computing timeline endYear and control highlights.

+++ web-app/backend/app/models/player.py
@@ -39,7 +39,7 @@ class Player(Base):
     avg_war = Column(Float)
     total_contract = Column(Float)
     avg_contract = Column(Float)
-    years_control = Column(Float)
+    years_control = Column(Integer)
     control_through = Column(Integer)
     total_future_war = Column(Float)
     total_future_value = Column(Float)
